requirements.txt
streamlit==1.35.0
pymongo==4.6.0
Pillow==10.1.0